requirements.txt
ansible>=2.9.0
pandas>=1.2.0
tabulate>=0.8.7
pyyaml>=5.4.1
jmespath>=0.10.0  # for json_query filter 